Handle NaN bit patterns in several RP copy stages
Many operations are implemented using float as the "working" format
within RP. It's easy for NaN bit patterns to not be preserved, causing
integer operations to get corrupted.

CL updates several of the existing tests to use NaN bit patterns (tests
previously failed x86-scalar), then includes fixes for them.

Reviewed-on: https://skia-review.googlesource.com/c/skia/+/799804
Commit-Queue: Brian Osman <brianosman@google.com>

+++ src/core/SkRasterPipelineOpContexts.h
@@ -243,8 +243,8 @@ struct SkRasterPipeline_SwizzleCopyCtx {
 };
 
 struct SkRasterPipeline_CopyIndirectCtx {
-    float* dst;
-    const float* src;
+    int32_t* dst;
+    const int32_t* src;
     const uint32_t *indirectOffset;  // this applies to `src` or `dst` based on the op
     uint32_t indirectLimit;          // the indirect offset is clamped to this upper bound
     uint32_t slots;                  // the number of slots to copy
